+++ src/stir_shaken_ssl.c
@@ -822,11 +822,11 @@ stir_shaken_status_t stir_shaken_x509_add_signalwire_extensions(stir_shaken_cont
     return STIR_SHAKEN_STATUS_OK;
 }
 
-stir_shaken_status_t stir_shaken_x509_req_add_tnauthlist_extension_spc(stir_shaken_context_t *ss, X509_REQ *req, int spc)
+stir_shaken_status_t stir_shaken_x509_req_add_tnauthlist_extension_spc(stir_shaken_context_t *ss, X509_REQ *req, unsigned long long int spc)
 {
     char buf[100] = { 0 };
 
-    snprintf(buf, 100, "%d", spc);
+    snprintf(buf, 100, "%llu", spc);
     return stir_shaken_v3_add_ext(NULL, NULL, req, NULL, stir_shaken_globals.tn_authlist_nid, buf);
 }
